moving things around and getting the orchestration service setup
