+++ app/src/main/java/com/ebsolutions/papertrail/financialdataingestionservice/accounttransaction/AccountTransactionController.java
@@ -1,10 +1,7 @@
 package com.ebsolutions.papertrail.financialdataingestionservice.accounttransaction;
 
-import com.ebsolutions.papertrail.financialdataingestionservice.accounttransaction.dto.AccountTransactionDto;
-import com.ebsolutions.papertrail.financialdataingestionservice.accounttransaction.factory.AccountTransactionFileReaderFactory;
-import com.ebsolutions.papertrail.financialdataingestionservice.accounttransaction.factory.AccountTransactionFileReaderFactoryRegistry;
+import com.ebsolutions.papertrail.financialdataingestionservice.accounttransaction.service.AccountTransactionIngestionOrchestrationService;
 import jakarta.validation.Valid;
-import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.MediaType;
@@ -22,23 +19,20 @@ import org.springframework.web.bind.annotation.RestController;
 @Slf4j
 public class AccountTransactionController {
 
-  private final AccountTransactionFileReaderFactoryRegistry factoryRegistry;
+  private final AccountTransactionIngestionOrchestrationService
+      accountTransactionIngestionOrchestrationService;
 
   @PostMapping(
       produces = MediaType.APPLICATION_JSON_VALUE,
       consumes = MediaType.MULTIPART_FORM_DATA_VALUE
   )
-  public ResponseEntity<?> loadFile(
+  public ResponseEntity<Void> loadFile(
       @ModelAttribute @Valid AccountTransactionFileEnvelope accountTransactionFileEnvelope)
       throws Exception {
 
-    AccountTransactionFileReaderFactory<? extends AccountTransactionDto> factory =
-        factoryRegistry.getFactory(accountTransactionFileEnvelope.getSupportedInstitution());
+    accountTransactionIngestionOrchestrationService.process(accountTransactionFileEnvelope);
 
-    AccountTransactionFileReaderService<? extends AccountTransactionDto> readerService =
-        factory.create(accountTransactionFileEnvelope.getSupportedInstitution());
-    List<? extends AccountTransactionDto> result =
-        readerService.process(accountTransactionFileEnvelope);
-    return ResponseEntity.ok(result);
+
+    return ResponseEntity.accepted().build();
   }
 }
