+++ app/src/main/java/com/ebsolutions/papertrail/financialdataingestionservice/accounttransaction/exception/AccountTransactionPublishException.java
@@ -1,4 +1,4 @@
-package com.ebsolutions.papertrail.financialdataingestionservice.accounttransaction;
+package com.ebsolutions.papertrail.financialdataingestionservice.accounttransaction.exception;
 
 public class AccountTransactionPublishException extends RuntimeException {
   public AccountTransactionPublishException(String message) {
